% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{abbreviate_class}
\alias{abbreviate_class}
\title{Abbreviate object class name}
\usage{
abbreviate_class(x, n = 4L)
}
\arguments{
\item{x}{Object.}

\item{n}{Integer: Minimum abbreviation length.}
}
\value{
Character: Abbreviated class wrapped in angle brackets.
}
\description{
Abbreviate object class name
}
\examples{
abbreviate_class(iris)
abbreviate_class(iris, n = 3)
}
\author{
EDG
}
