% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/03_S7_properties.R
\name{character_scalar}
\alias{character_scalar}
\title{Non-empty character scalar S7 property}
\usage{
character_scalar
}
\value{
An S7 property object.
}
\description{
S7 property accepting a single non-NA, non-empty (after trimming whitespace) string.
}
\examples{
Model <- S7::new_class("Model", properties = list(algorithm = character_scalar))
Model(algorithm = "LightGBM")@algorithm
try(Model(algorithm = ""))
}
\author{
EDG
}
