% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check.R
\name{check_character_scalar}
\alias{check_character_scalar}
\title{Check character scalar}
\usage{
check_character_scalar(x, arg_name = deparse(substitute(x)))
}
\arguments{
\item{x}{Character: Value to check. Must be a single non-NA, non-empty string.}

\item{arg_name}{Character: Argument name to use in error messages.}
}
\value{
Called for side effects. Throws an error if checks fail.
}
\description{
Check character scalar
}
\examples{
check_character_scalar("hello")
# Throw error:
try(check_character_scalar(""))
try(check_character_scalar(NA_character_))
try(check_character_scalar(c("a", "b")))
}
\author{
EDG
}
