% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check.R
\name{check_data.table}
\alias{check_data.table}
\title{Check data.table}
\usage{
check_data.table(x, arg_name = deparse(substitute(x)))
}
\arguments{
\item{x}{Object to check.}

\item{arg_name}{Character: Name of the variable for error messages.}
}
\value{
Called for side effects. Throws an error if input is not a data.table, returns x
invisibly otherwise.
}
\description{
Check data.table
}
\author{
EDG
}
