% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check.R
\name{check_double_scalar}
\alias{check_double_scalar}
\title{Check double scalar}
\usage{
check_double_scalar(x, arg_name = deparse(substitute(x)))
}
\arguments{
\item{x}{Numeric: Value to check. Must be a single non-NA number (integer inputs are accepted).}

\item{arg_name}{Character: Argument name to use in error messages.}
}
\value{
Called for side effects. Throws an error if checks fail.
}
\description{
Check double scalar
}
\examples{
check_double_scalar(3.14)
check_double_scalar(1L)
# Throw error:
try(check_double_scalar(NA_real_))
try(check_double_scalar(c(1.0, 2.0)))
}
\author{
EDG
}
