% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check.R
\name{check_enum}
\alias{check_enum}
\title{Check if value is in set of allowed values}
\usage{
check_enum(x, allowed_values, arg_name = deparse(substitute(x)))
}
\arguments{
\item{x}{Value to check.}

\item{allowed_values}{Vector of allowed values.}

\item{arg_name}{Character: Name of the variable for error messages.}
}
\value{
Called for side effects. Throws an error if x is not in allowed_values, returns x
invisibly otherwise.
}
\description{
Checks if a value is in a set of allowed values, and throws an error if not.
}
\examples{
check_enum("apple", c("apple", "banana", "cherry"))
# Throws error:
try(check_enum("granola", c("croissant", "bagel", "scramble")))
}
\author{
EDG
}
