% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check.R
\name{check_file_exists}
\alias{check_file_exists}
\title{Check file exists}
\usage{
check_file_exists(file)
}
\arguments{
\item{file}{Character: Path to file to check.}
}
\value{
Throws an error if checks fail. Returns the normalized file path invisibly if checks pass.
}
\description{
Check file exists
}
\examples{
\dontrun{
check_file_exists("path/to/file.txt")
}
}
\author{
EDG
}
