% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check.R
\name{check_float0pos}
\alias{check_float0pos}
\title{Check float greater than or equal to 0}
\usage{
check_float0pos(x, allow_null = TRUE, xname = deparse(substitute(x)))
}
\arguments{
\item{x}{Numeric vector.}

\item{allow_null}{Logical: If TRUE, NULL values are allowed and return early.}

\item{xname}{Character: Name of the variable for error messages.}
}
\value{
Called for side effects. Throws an error if checks fail.
}
\description{
Checks if an input is a numeric vector containing non-negative
(>= 0) values and no \code{NA}s. It is designed to validate function arguments.
}
\author{
EDG
}
