% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check.R
\name{check_floatpos}
\alias{check_floatpos}
\title{Check positive float}
\usage{
check_floatpos(x, allow_null = TRUE, arg_name = deparse(substitute(x)))
}
\arguments{
\item{x}{Numeric vector.}

\item{allow_null}{Logical: If TRUE, NULL values are allowed and return early.}

\item{arg_name}{Character: Name of the variable for error messages.}
}
\value{
Called for side effects. Throws an error if checks fail.
}
\description{
Check positive float
}
\details{
Checking with \code{is.numeric()} allows integer inputs as well, which should be ok since it is
unlikely the function that consumes this will enforce double type only, but instead is most
likely to allow implicit coercion from integer to numeric.
}
\examples{
check_floatpos(c(0.5, 1.5))
# Allows integers since they are numeric and can be coerced to double without loss of information
check_floatpos(c(1L, 3L))
# Throws error:
try(check_floatpos(c(-1.5, 0.5, 1.5)))
}
\author{
EDG
}
