% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check.R
\name{check_floatpos1}
\alias{check_floatpos1}
\title{Check float in (0, 1]}
\usage{
check_floatpos1(x, allow_null = TRUE, xname = deparse(substitute(x)))
}
\arguments{
\item{x}{Numeric vector.}

\item{allow_null}{Logical: If TRUE, NULL values are allowed and return early.}

\item{xname}{Character: Name of the variable for error messages.}
}
\value{
Called for side effects. Throws an error if checks fail.
}
\description{
Check float in (0, 1]
}
\author{
EDG
}
