% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check.R
\name{check_inherits}
\alias{check_inherits}
\title{Check class of object}
\usage{
check_inherits(x, cl, allow_null = TRUE, arg_name = deparse(substitute(x)))
}
\arguments{
\item{x}{Object to check.}

\item{cl}{Character: class to check against.}

\item{allow_null}{Logical: If TRUE, NULL values are allowed and return early.}

\item{arg_name}{Character: Name of the variable for error messages.}
}
\value{
Called for side effects. Throws an error if checks fail.
}
\description{
Check class of object
}
\examples{
check_inherits("papaya", "character")
# These will throw errors:
try(check_inherits(c(1, 2.5, 3.2), "integer"))
try(check_inherits(iris, "list"))
}
\author{
EDG
}
