% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check.R
\name{check_logical_scalar}
\alias{check_logical_scalar}
\title{Check logical scalar}
\usage{
check_logical_scalar(x, arg_name = deparse(substitute(x)))
}
\arguments{
\item{x}{Logical: Value to check. Must be a single non-NA \code{TRUE} or \code{FALSE}.}

\item{arg_name}{Character: Argument name to use in error messages.}
}
\value{
Called for side effects. Throws an error if checks fail.
}
\description{
Check logical scalar
}
\examples{
check_logical_scalar(TRUE)
check_logical_scalar(FALSE)
# Throw error:
try(check_logical_scalar(NA))
try(check_logical_scalar(1L))
try(check_logical_scalar(c(TRUE, FALSE)))
}
\author{
EDG
}
