% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check.R
\name{check_nonneg_double_scalar}
\alias{check_nonneg_double_scalar}
\title{Check non-negative double scalar}
\usage{
check_nonneg_double_scalar(x, arg_name = deparse(substitute(x)))
}
\arguments{
\item{x}{Numeric: Value to check. Must be a single finite number greater than or equal to zero.}

\item{arg_name}{Character: Argument name to use in error messages.}
}
\value{
Called for side effects. Throws an error if checks fail.
}
\description{
Check non-negative double scalar
}
\examples{
check_nonneg_double_scalar(0)
check_nonneg_double_scalar(5)
# Throw error:
try(check_nonneg_double_scalar(-0.001))
try(check_nonneg_double_scalar(Inf))
}
\author{
EDG
}
