% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check.R
\name{check_optional_bounded_integer_scalar}
\alias{check_optional_bounded_integer_scalar}
\title{Check optional integer scalar within bounds}
\usage{
check_optional_bounded_integer_scalar(
  x,
  lower = -Inf,
  upper = Inf,
  arg_name = deparse(substitute(x))
)
}
\arguments{
\item{x}{Optional Numeric: Value to check. Must be \code{NULL} or a single non-NA whole
number within bounds.}

\item{lower}{Numeric scalar: Lower bound, inclusive.}

\item{upper}{Numeric scalar: Upper bound, inclusive.}

\item{arg_name}{Character: Argument name to use in error messages.}
}
\value{
Called for side effects. Throws an error if checks fail.
}
\description{
Check optional integer scalar within bounds
}
\examples{
check_optional_bounded_integer_scalar(NULL)
check_optional_bounded_integer_scalar(5L, lower = 0, upper = 20)
# Throw error:
try(check_optional_bounded_integer_scalar(25, lower = 0, upper = 20))
}
\author{
EDG
}
