% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check.R
\name{check_optional_character_scalar}
\alias{check_optional_character_scalar}
\title{Check optional character scalar}
\usage{
check_optional_character_scalar(x, arg_name = deparse(substitute(x)))
}
\arguments{
\item{x}{Optional Character: Value to check. Must be \code{NULL} or a single non-NA, non-empty string.}

\item{arg_name}{Character: Argument name to use in error messages.}
}
\value{
Called for side effects. Throws an error if checks fail.
}
\description{
Check optional character scalar
}
\examples{
check_optional_character_scalar(NULL)
check_optional_character_scalar("hello")
# Throw error:
try(check_optional_character_scalar(""))
try(check_optional_character_scalar(c("a", "b")))
}
\author{
EDG
}
