% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check.R
\name{check_optional_integer_scalar}
\alias{check_optional_integer_scalar}
\title{Check optional integer scalar}
\usage{
check_optional_integer_scalar(x, arg_name = deparse(substitute(x)))
}
\arguments{
\item{x}{Optional Numeric: Value to check. Must be \code{NULL} or a single non-NA whole number.}

\item{arg_name}{Character: Argument name to use in error messages.}
}
\value{
Called for side effects. Throws an error if checks fail.
}
\description{
Check optional integer scalar
}
\examples{
check_optional_integer_scalar(NULL)
check_optional_integer_scalar(10L)
# Throw error:
try(check_optional_integer_scalar(1.5))
}
\author{
EDG
}
