% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check.R
\name{check_optional_logical_scalar}
\alias{check_optional_logical_scalar}
\title{Check optional logical scalar}
\usage{
check_optional_logical_scalar(x, arg_name = deparse(substitute(x)))
}
\arguments{
\item{x}{Optional Logical: Value to check. Must be \code{NULL} or a single non-NA \code{TRUE} or \code{FALSE}.}

\item{arg_name}{Character: Argument name to use in error messages.}
}
\value{
Called for side effects. Throws an error if checks fail.
}
\description{
Check optional logical scalar
}
\examples{
check_optional_logical_scalar(NULL)
check_optional_logical_scalar(FALSE)
# Throw error:
try(check_optional_logical_scalar(NA))
}
\author{
EDG
}
