% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check.R
\name{check_optional_nonneg_double_scalar}
\alias{check_optional_nonneg_double_scalar}
\title{Check optional non-negative double scalar}
\usage{
check_optional_nonneg_double_scalar(x, arg_name = deparse(substitute(x)))
}
\arguments{
\item{x}{Optional Numeric: Value to check. Must be \code{NULL} or a single finite number
greater than or equal to zero.}

\item{arg_name}{Character: Argument name to use in error messages.}
}
\value{
Called for side effects. Throws an error if checks fail.
}
\description{
Check optional non-negative double scalar
}
\examples{
check_optional_nonneg_double_scalar(NULL)
check_optional_nonneg_double_scalar(0)
# Throw error:
try(check_optional_nonneg_double_scalar(-1))
}
\author{
EDG
}
