% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check.R
\name{check_optional_nonneg_double_vector}
\alias{check_optional_nonneg_double_vector}
\title{Check optional non-negative double vector}
\usage{
check_optional_nonneg_double_vector(x, arg_name = deparse(substitute(x)))
}
\arguments{
\item{x}{Optional Numeric: Value to check. Must be \code{NULL} or a non-empty vector with all
elements finite, greater than or equal to zero, and no NAs.}

\item{arg_name}{Character: Argument name to use in error messages.}
}
\value{
Called for side effects. Throws an error if checks fail.
}
\description{
Check optional non-negative double vector
}
\examples{
check_optional_nonneg_double_vector(NULL)
check_optional_nonneg_double_vector(c(0, 1, 5))
# Throw error:
try(check_optional_nonneg_double_vector(c(-1, 0)))
}
\author{
EDG
}
