% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check.R
\name{check_optional_prob_scalar}
\alias{check_optional_prob_scalar}
\title{Check optional probability scalar}
\usage{
check_optional_prob_scalar(x, arg_name = deparse(substitute(x)))
}
\arguments{
\item{x}{Optional Numeric: Value to check. Must be \code{NULL} or a single finite number in \eqn{[0, 1]}.}

\item{arg_name}{Character: Argument name to use in error messages.}
}
\value{
Called for side effects. Throws an error if checks fail.
}
\description{
Check optional probability scalar
}
\examples{
check_optional_prob_scalar(NULL)
check_optional_prob_scalar(0.5)
# Throw error:
try(check_optional_prob_scalar(2.0))
}
\author{
EDG
}
