% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check.R
\name{check_optional_scalar_character}
\alias{check_optional_scalar_character}
\title{Check Optional Scalar Character}
\usage{
check_optional_scalar_character(x, arg_name = deparse(substitute(x)))
}
\arguments{
\item{x}{Optional Character: Value to check.}

\item{arg_name}{Character: Argument name to use in error messages.}
}
\value{
Called for side effects.
}
\description{
Check Optional Scalar Character
}
\examples{
check_optional_scalar_character(NULL, "my_arg")
check_optional_scalar_character("hello", "my_arg")
# Throw error:
try(check_optional_scalar_character(c("hello", "world"), "my_arg"))
try(check_optional_scalar_character(123, "my_arg"))
}
\author{
EDG
}
