% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check.R
\name{check_optional_unit_open_vector}
\alias{check_optional_unit_open_vector}
\title{Check optional open-unit-interval vector}
\usage{
check_optional_unit_open_vector(x, arg_name = deparse(substitute(x)))
}
\arguments{
\item{x}{Optional Numeric: Value to check. Must be \code{NULL} or a non-empty vector with all
elements strictly in \eqn{(0, 1)} and no NAs.}

\item{arg_name}{Character: Argument name to use in error messages.}
}
\value{
Called for side effects. Throws an error if checks fail.
}
\description{
Check optional open-unit-interval vector
}
\examples{
check_optional_unit_open_vector(NULL)
check_optional_unit_open_vector(c(0.1, 0.9))
# Throw error:
try(check_optional_unit_open_vector(c(0, 0.5)))
}
\author{
EDG
}
