% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check.R
\name{check_pos_double_vector}
\alias{check_pos_double_vector}
\title{Check positive double vector}
\usage{
check_pos_double_vector(x, arg_name = deparse(substitute(x)))
}
\arguments{
\item{x}{Numeric: Value to check. Must be a non-empty vector with all elements finite,
strictly greater than zero, and no NAs.}

\item{arg_name}{Character: Argument name to use in error messages.}
}
\value{
Called for side effects. Throws an error if checks fail.
}
\description{
Check positive double vector
}
\examples{
check_pos_double_vector(c(0.1, 1, 10))
# Throw error:
try(check_pos_double_vector(c(0, 1)))
try(check_pos_double_vector(c(1, Inf)))
}
\author{
EDG
}
