% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check.R
\name{check_prob_scalar}
\alias{check_prob_scalar}
\title{Check probability scalar}
\usage{
check_prob_scalar(x, arg_name = deparse(substitute(x)))
}
\arguments{
\item{x}{Numeric: Value to check. Must be a single finite number in \eqn{[0, 1]}.}

\item{arg_name}{Character: Argument name to use in error messages.}
}
\value{
Called for side effects. Throws an error if checks fail.
}
\description{
Check probability scalar
}
\examples{
check_prob_scalar(0)
check_prob_scalar(0.5)
check_prob_scalar(1)
# Throw error:
try(check_prob_scalar(1.5))
try(check_prob_scalar(-0.1))
}
\author{
EDG
}
