% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check.R
\name{check_prob_vector}
\alias{check_prob_vector}
\title{Check probability vector}
\usage{
check_prob_vector(x, arg_name = deparse(substitute(x)))
}
\arguments{
\item{x}{Numeric: Value to check. Must be a non-empty vector with all elements in \eqn{[0, 1]}
and no NAs.}

\item{arg_name}{Character: Argument name to use in error messages.}
}
\value{
Called for side effects. Throws an error if checks fail.
}
\description{
Check probability vector
}
\examples{
check_prob_vector(c(0, 0.5, 1))
# Throw error:
try(check_prob_vector(c(0.5, 1.5)))
try(check_prob_vector(c(0.5, NA)))
}
\author{
EDG
}
