% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check.R
\name{check_scalar_logical}
\alias{check_scalar_logical}
\title{Check Scalar Logical}
\usage{
check_scalar_logical(x, arg_name)
}
\arguments{
\item{x}{Logical: Value to check.}

\item{arg_name}{Character: Argument name to use in error messages.}
}
\value{
Called for side effects.
}
\description{
Check Scalar Logical
}
\examples{
check_scalar_logical(TRUE, "my_arg") # Passes
check_scalar_logical(FALSE, "my_arg") # Passes
# Throw error:
try(check_scalar_logical(c(TRUE, FALSE), "my_arg"))
try(check_scalar_logical(NA, "my_arg"))
try(check_scalar_logical("TRUE", "my_arg"))
}
\author{
EDG
}
