% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check.R
\name{check_tabular}
\alias{check_tabular}
\title{Check object is tabular}
\usage{
check_tabular(x)
}
\arguments{
\item{x}{Object to check.}
}
\value{
Called for side effects. Throws an error if input is not tabular, returns x invisibly
otherwise.
}
\description{
Checks if object is of class \code{data.frame}, \code{data.table}, or \code{tbl_df}.
}
\examples{
check_tabular(iris)
check_tabular(data.table::as.data.table(iris))
# Throws error:
try(check_tabular(matrix(1:10, ncol = 2)))
}
\author{
EDG
}
