% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check.R
\name{check_unit_open_scalar}
\alias{check_unit_open_scalar}
\title{Check open-unit-interval scalar}
\usage{
check_unit_open_scalar(x, arg_name = deparse(substitute(x)))
}
\arguments{
\item{x}{Numeric: Value to check. Must be a single finite number strictly in \eqn{(0, 1)}.}

\item{arg_name}{Character: Argument name to use in error messages.}
}
\value{
Called for side effects. Throws an error if checks fail.
}
\description{
Check open-unit-interval scalar
}
\examples{
check_unit_open_scalar(0.5)
# Throw error:
try(check_unit_open_scalar(0))
try(check_unit_open_scalar(1))
}
\author{
EDG
}
