% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_strings.R
\name{checkmark}
\alias{checkmark}
\title{Checkmark}
\usage{
checkmark(col = col_success, output_type = c("ansi", "html", "plain"))
}
\arguments{
\item{col}{Color for the checkmark symbol.}

\item{output_type}{Character \{"ansi", "html", "plain"\}: Output type for formatting.}
}
\value{
Character: Formatted checkmark symbol.
}
\description{
Prints a checkmark symbol with optional color and formatting.
}
\examples{
checkmark()
}
\author{
EDG
}
