% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean.R
\name{clean_posint}
\alias{clean_posint}
\title{Check positive integer}
\usage{
clean_posint(x, allow_na = FALSE, arg_name = deparse(substitute(x)))
}
\arguments{
\item{x}{Integer vector.}

\item{allow_na}{Logical: If TRUE, NAs are excluded before checking. If FALSE (default),
NAs trigger an error.}

\item{arg_name}{Character: Name of the variable for error messages.}
}
\value{
Integer vector of positive values.
}
\description{
Check positive integer
}
\examples{
clean_posint(5)
}
\author{
EDG
}
