% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_print.R
\name{collapse_head}
\alias{collapse_head}
\title{Collapse head of vector with commas followed by ellipsis}
\usage{
collapse_head(x, maxlength = 6L, format_fn = identity)
}
\arguments{
\item{x}{Vector: Input whose first elements are shown.}

\item{maxlength}{Integer: Maximum number of elements to show before truncating with an ellipsis. Use -1 to show all.}

\item{format_fn}{Function: Formatting function applied to each element.}
}
\value{
Character.
}
\description{
Collapse head of vector with commas followed by ellipsis
}
\details{
Used, for example, by \code{repr_ls}
}
\examples{
collapse_head(98054:99890, maxlength = 5L)
collapse_head(
  c("mango", "banana", "tangerine", "sugar", "ackee", "cocoa bean"),
  maxlength = 3L, format_fn = toupper
)
}
\author{
EDG
}
