% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_strings.R
\name{crossmark}
\alias{crossmark}
\title{Cross mark}
\usage{
crossmark(col = rtemis_colors[["red"]], output_type = NULL)
}
\arguments{
\item{col}{Color for the cross mark symbol.}

\item{output_type}{Character ("ansi", "html", "plain") or NULL: Output type. If NULL, resolved
via \code{\link[=get_output_type]{get_output_type()}}.}
}
\value{
Character: Formatted cross mark symbol.
}
\description{
Prints a cross mark symbol with optional color and formatting.
}
\examples{
crossmark()
}
\author{
EDG
}
