% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/msg.R
\name{datetime}
\alias{datetime}
\title{Get current date and time}
\usage{
datetime(datetime_format = "\%Y-\%m-\%d \%H:\%M:\%S")
}
\arguments{
\item{datetime_format}{Character: Format for the date and time.}
}
\value{
Character: Formatted date and time.
}
\description{
Used by \code{msgdatetime()} and \code{log_to_file()}.
}
\examples{
datetime()
}
\author{
EDG
}
\keyword{internal}
