% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fmt.R
\name{fmt_gradient}
\alias{fmt_gradient}
\title{Gradient text}
\usage{
fmt_gradient(x, colors, bold = FALSE, output_type = NULL)
}
\arguments{
\item{x}{Character: Text to colorize.}

\item{colors}{Character vector: Colors to use for the gradient.}

\item{bold}{Logical: If TRUE, make text bold.}

\item{output_type}{Character ("ansi", "html", "plain") or NULL: Output type. If NULL, resolved
via \code{\link[=get_output_type]{get_output_type()}}.}
}
\value{
Character: Text with gradient color applied.
}
\description{
Gradient text
}
\examples{
fmt_gradient("Gradient Text", colors = c("blue", "red")) |> message()
}
\author{
EDG
}
