% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_strings.R
\name{fn2label}
\alias{fn2label}
\title{Function to label}
\usage{
fn2label(fn, varname)
}
\arguments{
\item{fn}{Function.}

\item{varname}{Character: Variable name.}
}
\value{
Character: Label.
}
\description{
Create axis label from function definition and variable name
}
\examples{
fn2label(mean, "age")
}
\author{
EDG
}
\keyword{internal}
