% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fmt.R
\name{gray}
\alias{gray}
\title{Gray text}
\usage{
gray(x, output_type = NULL)
}
\arguments{
\item{x}{Character: Text to format}

\item{output_type}{Character ("ansi", "html", "plain") or NULL: Output type. If NULL, resolved
via \code{\link[=get_output_type]{get_output_type()}}.}
}
\value{
Character: Formatted text with gray styling
}
\description{
A \code{fmt()} convenience wrapper for making text gray.
}
\details{
This is an internal function so it does not mask \code{grDevices::gray()}.
Can be useful in contexts where muted is not supported.
}
\examples{
message(gray("gray text"))
}
\author{
EDG
}
\keyword{internal}
