% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fmt.R
\name{highlight}
\alias{highlight}
\title{Highlight text}
\usage{
highlight(x, pad = 0L, output_type = NULL)
}
\arguments{
\item{x}{Character: Text to highlight.}

\item{pad}{Integer: Number of spaces to pad before text.}

\item{output_type}{Character ("ansi", "html", "plain") or NULL: Output type. If NULL, resolved
via \code{\link[=get_output_type]{get_output_type()}}.}
}
\value{
Character: Formatted text with highlight.
}
\description{
A \code{fmt()} convenience wrapper for highlighting text.
}
\examples{
message(highlight("This is highlighted!"))
}
\author{
EDG
}
