% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_strings.R
\name{highlightbig}
\alias{highlightbig}
\title{Highlight big numbers}
\usage{
highlightbig(x, output_type = NULL)
}
\arguments{
\item{x}{Numeric: Input}

\item{output_type}{Character ("ansi", "html", "plain") or NULL: Output type. If NULL, resolved
via \code{\link[=get_output_type]{get_output_type()}}.}
}
\value{
Character: Formatted number with thousands separators and highlighting.
}
\description{
Highlight big numbers
}
\examples{
message(highlightbig(1234567))
}
\keyword{internal}
