% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/html.R
\name{html_escape}
\alias{html_escape}
\title{Escape text for inclusion in HTML}
\usage{
html_escape(x, attribute = FALSE)
}
\arguments{
\item{x}{Character: Text to escape. Coerced with \code{as.character()}.}

\item{attribute}{Logical: If TRUE, also escape the quote characters, which
must be escaped inside an attribute value but not in element text. Must be a
single non-NA TRUE or FALSE.}
}
\value{
Character: Escaped text, carrying no class.
}
\description{
Escape text for inclusion in HTML
}
\examples{
html_escape("a < b & c")
html_escape('say "hi"', attribute = TRUE)
}
\author{
EDG
}
