% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/html.R
\name{html_p}
\alias{html_p}
\title{Build a \code{p} element}
\usage{
html_p(..., class = NULL, style = NULL, id = NULL)
}
\arguments{
\item{...}{Children: Character, numbers, \code{rtemis_html} objects, or lists of
those.}

\item{class}{Optional Character: \code{class} attribute. Must be NULL or a single
string; collapse multiple classes yourself, e.g.
\code{paste(x, collapse = " ")}.}

\item{style}{Optional Character: \code{style} attribute. Must be NULL or a single
string.}

\item{id}{Optional Character: \code{id} attribute. Must be NULL or a single
string.}
}
\value{
Character of class \code{rtemis_html_element} and \code{rtemis_html}.
}
\description{
Build a \code{p} element
}
\examples{
html_p("A paragraph.")
}
\author{
EDG
}
