% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/03_S7_properties.R
\name{logical_scalar}
\alias{logical_scalar}
\title{Logical scalar S7 property}
\usage{
logical_scalar
}
\value{
An S7 property object.
}
\description{
S7 property accepting a single non-NA logical value.
}
\author{
EDG
}
