% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{match_arg}
\alias{match_arg}
\title{Match Arguments Ignoring Case}
\usage{
match_arg(x, choices)
}
\arguments{
\item{x}{Character: Argument to match.}

\item{choices}{Character vector: Choices to match against.}
}
\value{
Character: Matched argument.
}
\description{
Match Arguments Ignoring Case
}
\examples{
match_arg("papaya", c("AppleExtreme", "SuperBanana", "PapayaMaster"))
}
\author{
EDG
}
