% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/msg.R
\name{msg}
\alias{msg}
\alias{msg0}
\title{Message with provenance}
\usage{
msg(
  ...,
  caller = NULL,
  call_depth = 1L,
  caller_id = 1L,
  newline_pre = FALSE,
  newline = TRUE,
  format_fn = plain,
  sep = " ",
  verbosity = 1L
)

msg0(
  ...,
  caller = NULL,
  call_depth = 1,
  caller_id = 1,
  newline_pre = FALSE,
  newline = TRUE,
  format_fn = plain,
  sep = "",
  verbosity = 1L
)
}
\arguments{
\item{...}{Message to print}

\item{caller}{Character: Name of calling function}

\item{call_depth}{Integer: Print the system call path of this depth.}

\item{caller_id}{Integer: Which function in the call stack to print}

\item{newline_pre}{Logical: If TRUE begin with a new line.}

\item{newline}{Logical: If TRUE end with a new line.}

\item{format_fn}{Function: Formatting function to use on the message text.}

\item{sep}{Character: Use to separate objects in \code{...}}

\item{verbosity}{Integer: Verbosity level of the message. If 0L, does not print anything and
returns NULL, invisibly.}
}
\value{
If verbosity > 0L, returns a list with call, message, and date, invisibly, otherwise
returns NULL invisibly.
}
\description{
Print message to output with a prefix including data and time, and calling function or full
call stack
}
\details{
If \code{msg} is called directly from the console, it will print \verb{[interactive>]} in place of
the call stack.
\code{msg0}, similar to \code{paste0}, is \code{msg(..., sep = "")}
}
\examples{
msg("Hello")
}
\author{
EDG
}
