% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_strings.R
\name{nay}
\alias{nay}
\title{Failure message}
\usage{
nay(..., sep = " ", end = "\\n", pad = 0)
}
\arguments{
\item{...}{Character: Message components.}

\item{sep}{Character: Separator between message components.}

\item{end}{Character: End character.}

\item{pad}{Integer: Number of spaces to pad the message with.}
}
\value{
\code{NULL} invisibly; prints a failure message to the console.
}
\description{
Failure message
}
\examples{
nay("Operation failed")
}
\author{
EDG
}
\keyword{internal}
