% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/03_S7_properties.R
\name{nonneg_double_vector}
\alias{nonneg_double_vector}
\title{Non-negative double vector S7 property}
\usage{
nonneg_double_vector
}
\value{
An S7 property object.
}
\description{
S7 property accepting a non-empty double vector with all elements finite, greater than or
equal to zero, and no NAs.
}
\author{
EDG
}
