% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/03_S7_properties.R
\name{optional}
\alias{optional}
\title{Create an optional S7 type}
\usage{
optional(type)
}
\arguments{
\item{type}{S7 base class or S7 class.}
}
\value{
An S7 union type that allows for the specified type or \code{NULL}.
}
\description{
Creates an S7 union type that allows for the specified type or \code{NULL}.
}
\details{
This should be used when the S7 class already includes all the necessary validation for the
non-NULL case. Otherwise, create a new S7 property with appropriate validator using
\code{S7::new_property()}.
}
\examples{
# Create an optional character type
optional(S7::class_character)
}
\author{
EDG
}
