% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/03_S7_properties.R
\name{optional_nonneg_double_scalar}
\alias{optional_nonneg_double_scalar}
\title{Optional non-negative double scalar S7 property}
\usage{
optional_nonneg_double_scalar
}
\value{
An S7 property object.
}
\description{
S7 property accepting \code{NULL} or a single finite double greater than or equal to zero.
}
\author{
EDG
}
