% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/03_S7_properties.R
\name{optional_nonneg_double_vector}
\alias{optional_nonneg_double_vector}
\title{Optional non-negative double vector S7 property}
\usage{
optional_nonneg_double_vector
}
\value{
An S7 property object.
}
\description{
S7 property accepting \code{NULL} or a non-empty double vector with all elements finite,
greater than or equal to zero, and no NAs.
}
\examples{
Importance <- S7::new_class(
  "Importance",
  properties = list(scores = optional_nonneg_double_vector)
)
Importance()@scores
Importance(scores = c(0, 2))@scores
try(Importance(scores = c(0, -2)))
}
\author{
EDG
}
