% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/03_S7_properties.R
\name{optional_pos_double_scalar}
\alias{optional_pos_double_scalar}
\title{Optional positive double scalar S7 property}
\usage{
optional_pos_double_scalar
}
\value{
An S7 property object.
}
\description{
S7 property accepting \code{NULL} or a single finite double strictly greater than zero.
}
\author{
EDG
}
