% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/03_S7_properties.R
\name{optional_prob_vector}
\alias{optional_prob_vector}
\title{Optional probability vector S7 property}
\usage{
optional_prob_vector
}
\value{
An S7 property object.
}
\description{
S7 property accepting \code{NULL} or a non-empty double vector with all elements in \eqn{[0, 1]}
and no NAs.
}
\author{
EDG
}
