% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/03_S7_properties.R
\name{optional_unit_open_vector}
\alias{optional_unit_open_vector}
\title{Optional open-unit-interval vector S7 property}
\usage{
optional_unit_open_vector
}
\value{
An S7 property object.
}
\description{
S7 property accepting \code{NULL} or a non-empty double vector with all elements strictly in
\eqn{(0, 1)} and no NAs.
}
\examples{
Grid <- S7::new_class(
  "Grid",
  properties = list(subsamples = optional_unit_open_vector)
)
Grid()@subsamples
Grid(subsamples = c(0.5, 0.8))@subsamples
try(Grid(subsamples = c(0.5, 1)))
}
\author{
EDG
}
