% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_strings.R
\name{oxfordcomma}
\alias{oxfordcomma}
\title{Oxford comma}
\usage{
oxfordcomma(..., format_fn = identity)
}
\arguments{
\item{...}{Character vector: Items to be combined.}

\item{format_fn}{Function: Any function to be applied to each item.}
}
\value{
Character: Formatted string with oxford comma.
}
\description{
Oxford comma
}
\examples{
oxfordcomma("a", "b", "c")
}
\author{
EDG
}
\keyword{internal}
