% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/msg.R
\name{pad_string}
\alias{pad_string}
\title{Left-pad a string to a target width}
\usage{
pad_string(x, target = 17, char = " ")
}
\arguments{
\item{x}{Character: String to pad.}

\item{target}{Integer: Target total width.}

\item{char}{Character: Padding character.}
}
\value{
Character: \code{x} left-padded with \code{char} to width \code{target}.
}
\description{
Left-pad a string to a target width
}
\examples{
pad_string("hi", target = 6L)
}
\author{
EDG
}
\keyword{internal}
