% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/02_S7_prop_factories.R
\name{prop_boolean}
\alias{prop_boolean}
\title{Logical (boolean) S7 property}
\usage{
prop_boolean(default = FALSE, nullable = FALSE, description = "")
}
\arguments{
\item{default}{Logical: Default value (NULL only if \code{nullable}).}

\item{nullable}{Logical: If TRUE, NULL is a valid value.}

\item{description}{Character: Human-readable description.}
}
\value{
S7 property.
}
\description{
Logical (boolean) S7 property
}
\examples{
verbose <- prop_boolean(default = TRUE, description = "Print progress")
}
\author{
EDG
}
