% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_strings.R
\name{red}
\alias{red}
\title{Red}
\usage{
red(..., bold = FALSE)
}
\arguments{
\item{...}{Character: Text to colorize.}

\item{bold}{Logical: If TRUE, make text bold.}
}
\value{
Character: Red-colored text.
}
\description{
Red
}
\examples{
message(red("error"))
}
\author{
EDG
}
\keyword{internal}
