% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/04_S7_generics.R
\name{repr}
\alias{repr}
\title{String representation}
\usage{
repr(x, ...)
}
\arguments{
\item{x}{Object to represent as a string.}

\item{...}{Additional arguments passed to methods.}
}
\value{
Character string representation of the object.
}
\description{
String representation
}
\examples{
S7::method(repr, S7::class_character) <- function(x, ...) {
  paste0("<chr> \"", x, "\"")
}
cat(repr("hello"))
}
\author{
EDG
}
