% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_print.R
\name{repr_ls}
\alias{repr_ls}
\title{Show list as formatted string}
\usage{
repr_ls(
  x,
  prefix = "",
  pad = 2L,
  item_format = bold,
  maxlength = 4L,
  center_title = TRUE,
  title = NULL,
  title_newline = TRUE,
  newline_pre = FALSE,
  format_fn_rhs = ddSci,
  print_class = TRUE,
  abbrev_class_n = 3L,
  print_df = FALSE,
  print_S4 = FALSE,
  limit = 12L,
  output_type = NULL
)
}
\arguments{
\item{x}{list or object that will be converted to a list.}

\item{prefix}{Character: Optional prefix for names.}

\item{pad}{Integer: Pad output with this many spaces.}

\item{item_format}{Formatting function for items.}

\item{maxlength}{Integer: Maximum length of items to show using \code{collapse_head()} before truncating with ellipsis.}

\item{center_title}{Logical: If TRUE, autopad title for centering, if present.}

\item{title}{Character: Title to print before list.}

\item{title_newline}{Logical: If TRUE, print title on new line.}

\item{newline_pre}{Logical: If TRUE, print newline before list.}

\item{format_fn_rhs}{Formatting function for right-hand side of items.}

\item{print_class}{Logical: If TRUE, print abbreviated class of object.}

\item{abbrev_class_n}{Integer: Number of characters to abbreviate class names to.}

\item{print_df}{Logical: If TRUE, print data frame contents, otherwise print n rows and columns.}

\item{print_S4}{Logical: If TRUE, print S4 object contents, otherwise print class name.}

\item{limit}{Integer: Maximum number of items to show.}

\item{output_type}{Character ("ansi", "html", "plain") or NULL: Output type for mformat. If
NULL, resolved via \code{\link[=get_output_type]{get_output_type()}}.}
}
\value{
Character: Formatted string that can be printed with cat()
}
\description{
Works exactly like printls, but instead of printing to console with cat,
it outputs a single string, formatted using mformat, so that cat(repr_ls(x))
looks identical to printls(x) for any list x
}
\details{
Exported as internal function for use by other rtemis packages.
}
\examples{
x <- list(
  a = 1:10,
  b = "Hello",
  c = list(
    d = 1,
    e = 2
  )
)
cat(repr_ls(x, title = "A List"))
}
\author{
EDG
}
