% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_strings.R
\name{repr_S7name}
\alias{repr_S7name}
\title{Show S7 class name}
\usage{
repr_S7name(x, col = col_object, pad = 0L, prefix = NULL, output_type = NULL)
}
\arguments{
\item{x}{Character: S7 class name.}

\item{col}{Color: Color code for the object name.}

\item{pad}{Integer: Number of spaces to pad the message with.}

\item{prefix}{Character: Prefix to add to the object name.}

\item{output_type}{Character ("ansi", "html", "plain") or NULL: Output type. If NULL, resolved
via \code{\link[=get_output_type]{get_output_type()}}.}
}
\value{
Character: Formatted string that can be printed with cat().
}
\description{
Show S7 class name
}
\examples{
repr_S7name("Supervised") |> cat()
}
\author{
EDG
}
\keyword{internal}
