% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/core-package.R
\docType{package}
\name{rtemis.core-package}
\alias{rtemis.core}
\alias{rtemis.core-package}
\title{\pkg{rtemis.core}: Rtemis Utilities}
\description{
Core Utilities for rtemis R Packages
}
\author{
\strong{Maintainer}: E.D. Gennatas \email{gennatas@gmail.com} (\href{https://orcid.org/0000-0001-9280-3609}{ORCID}) [copyright holder]

}
