% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test.R
\name{test_inherits}
\alias{test_inherits}
\title{Check class of object}
\usage{
test_inherits(x, cl)
}
\arguments{
\item{x}{Object to check}

\item{cl}{Character: class to check against}
}
\value{
Logical
}
\description{
Check class of object
}
\examples{
test_inherits("papaya", "character") # TRUE
test_inherits(c(1, 2.5, 3.2), "integer")
test_inherits(iris, "list") # FALSE, compare to is_check(iris, is.list)
}
\author{
EDG
}
