% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fmt.R
\name{thin}
\alias{thin}
\title{Make text thin/light}
\usage{
thin(text, output_type = NULL)
}
\arguments{
\item{text}{Character: Text to make thin}

\item{output_type}{Character ("ansi", "html", "plain") or NULL: Output type. If NULL, resolved
via \code{\link[=get_output_type]{get_output_type()}}.}
}
\value{
Character: Formatted text with thin/light styling
}
\description{
A \code{fmt()} convenience wrapper for making text thin/light.
}
\examples{
message(thin("thin text"))
}
\author{
EDG
}
\keyword{internal}
