% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/03_S7_properties.R
\name{unit_open_scalar}
\alias{unit_open_scalar}
\title{Open-unit-interval scalar S7 property}
\usage{
unit_open_scalar
}
\value{
An S7 property object.
}
\description{
S7 property accepting a single finite double strictly in \eqn{(0, 1)}.
}
\examples{
Boost <- S7::new_class("Boost", properties = list(learning_rate = unit_open_scalar))
Boost(learning_rate = 0.1)@learning_rate
# Bounds are exclusive: 0 and 1 are rejected
try(Boost(learning_rate = 1))
}
\author{
EDG
}
