% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/schema.R
\name{write_lines}
\alias{write_lines}
\title{Write lines to file}
\usage{
write_lines(x, file, overwrite = FALSE, verbosity = 1L)
}
\arguments{
\item{x}{Character: Text to write to file.}

\item{file}{Character: Path to output file.}

\item{overwrite}{Logical: If TRUE, overwrite an existing file.}

\item{verbosity}{Integer: Verbosity level.}
}
\value{
Invisible NULL. Called for the side effect of writing to file.
}
\description{
Normalizes path, check if directory exists, creates it if necessary,
writes lines to file, and checks if file was created successfully.
}
\examples{
tmpfile <- file.path(tempdir(), "demo.txt")
write_lines("hello", tmpfile, overwrite = TRUE, verbosity = 0L)
}
\author{
EDG
}
