% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_strings.R
\name{yay}
\alias{yay}
\title{Success message}
\usage{
yay(..., sep = " ", end = "\\n", pad = 0)
}
\arguments{
\item{...}{Character: Message components.}

\item{sep}{Character: Separator between message components.}

\item{end}{Character: End character.}

\item{pad}{Integer: Number of spaces to pad the message with.}
}
\value{
\code{NULL} invisibly; prints a success message to the console.
}
\description{
Success message
}
\examples{
yay("Operation complete")
}
\author{
EDG
}
\keyword{internal}
