% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fmt.R
\name{bold}
\alias{bold}
\title{Make text bold}
\usage{
bold(text, reset_code = "\\033[22m", output_type = c("ansi", "html", "plain"))
}
\arguments{
\item{text}{Character: Text to make bold.}

\item{reset_code}{Character: ANSI reset code to use after formatting.}

\item{output_type}{Character: Output type ("ansi", "html", "plain").}
}
\value{
Character: Formatted text with bold styling
}
\description{
A \code{fmt()} convenience wrapper for making text bold.
}
\examples{
message(bold("This is bold!"))
}
\author{
EDG
}
