% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/00_S7_properties.R
\docType{data}
\name{character_scalar}
\alias{character_scalar}
\title{Non-empty character scalar S7 property}
\format{
An object of class \code{S7_property} of length 6.
}
\usage{
character_scalar
}
\value{
An S7 property object.
}
\description{
S7 property accepting a single non-NA, non-empty (after trimming whitespace) string.
}
\author{
EDG
}
\keyword{datasets}
