% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check.R
\name{check_float0pos}
\alias{check_float0pos}
\title{Check float greater than or equal to 0}
\usage{
check_float0pos(x, allow_null = TRUE, arg_name = deparse(substitute(x)))
}
\arguments{
\item{x}{Numeric vector.}

\item{allow_null}{Logical: If TRUE, NULL values are allowed and return early.}

\item{arg_name}{Character: Name of the variable for error messages.}
}
\value{
Called for side effects. Throws an error if checks fail.
}
\description{
Checks if an input is a numeric vector containing non-negative
(>= 0) values and no \code{NA}s. It is designed to validate function arguments.
}
\examples{
check_float0pos(c(0, 0.5, 1))
# Allows integers since they are numeric and can be coerced to double without loss of information
check_float0pos(c(0L, 1L))
# Throws error:
try(check_float0pos(c(-1.5, 0, 1.5)))
}
\author{
EDG
}
