% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check.R
\name{check_logical}
\alias{check_logical}
\title{Check logical}
\usage{
check_logical(x, allow_null = TRUE, arg_name = deparse(substitute(x)))
}
\arguments{
\item{x}{Vector to check.}

\item{allow_null}{Logical: If TRUE, NULL values are allowed and return early.}

\item{arg_name}{Character: Name of the variable for error messages.}
}
\value{
Called for side effects. Throws an error if checks fail.
}
\description{
Check logical
}
\examples{
check_logical(c(TRUE, FALSE))
# Throws error:
try(check_logical(c(0, 1)))
}
\author{
EDG
}
