% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check.R
\name{check_optional_bounded_double_scalar}
\alias{check_optional_bounded_double_scalar}
\title{Check optional double scalar within bounds}
\usage{
check_optional_bounded_double_scalar(
  x,
  lower = -Inf,
  upper = Inf,
  lower_open = FALSE,
  upper_open = FALSE,
  arg_name = deparse(substitute(x))
)
}
\arguments{
\item{x}{Optional Numeric: Value to check. Must be \code{NULL} or a single non-NA number
within bounds.}

\item{lower}{Numeric scalar: Lower bound. Default \code{-Inf}.}

\item{upper}{Numeric scalar: Upper bound. Default \code{Inf}.}

\item{lower_open}{Logical scalar: If \code{TRUE}, the lower bound is exclusive. Default
\code{FALSE}.}

\item{upper_open}{Logical scalar: If \code{TRUE}, the upper bound is exclusive. Default
\code{FALSE}.}

\item{arg_name}{Character: Argument name to use in error messages.}
}
\value{
Called for side effects. Throws an error if checks fail.
}
\description{
Check optional double scalar within bounds
}
\examples{
check_optional_bounded_double_scalar(NULL)
check_optional_bounded_double_scalar(0.7, lower = 0, upper = 2)
# Throw error:
try(check_optional_bounded_double_scalar(5, lower = 0, upper = 2))
}
\author{
EDG
}
