% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_strings.R
\name{checkmark}
\alias{checkmark}
\title{Checkmark}
\usage{
checkmark(col = col_success, output_type = NULL)
}
\arguments{
\item{col}{Color for the checkmark symbol.}

\item{output_type}{Character ("ansi", "html", "plain") or NULL: Output type. If NULL, resolved
via \code{\link[=get_output_type]{get_output_type()}}.}
}
\value{
Character: Formatted checkmark symbol.
}
\description{
Prints a checkmark symbol with optional color and formatting.
}
\examples{
checkmark()
}
\author{
EDG
}
