% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_strings.R
\name{crossmark}
\alias{crossmark}
\title{Cross mark}
\usage{
crossmark(
  col = rtemis_colors[["red"]],
  output_type = c("ansi", "html", "plain")
)
}
\arguments{
\item{col}{Color for the cross mark symbol.}

\item{output_type}{Character \{"ansi", "html", "plain"\}: Output type for formatting.}
}
\value{
Character: Formatted cross mark symbol.
}
\description{
Prints a cross mark symbol with optional color and formatting.
}
\examples{
crossmark()
}
\author{
EDG
}
