% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/03_S7_properties.R
\name{double_scalar}
\alias{double_scalar}
\title{Double scalar S7 property}
\usage{
double_scalar
}
\value{
An S7 property object.
}
\description{
S7 property accepting a single non-NA double value.
}
\examples{
Metric <- S7::new_class("Metric", properties = list(value = double_scalar))
Metric(value = -1.5)@value
try(Metric(value = c(1, 2)))
}
\author{
EDG
}
