% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/00_S7_properties.R
\name{enum}
\alias{enum}
\title{Create an enum S7 property}
\usage{
enum(values, default = NULL, nullable = FALSE)
}
\arguments{
\item{values}{Character: Allowed values.}

\item{default}{Optional Character: Default value.}

\item{nullable}{Logical scalar. If \code{TRUE}, \code{NULL} is also accepted. Default \code{FALSE}.}
}
\value{
An S7 property object.
}
\description{
Returns a \code{new_property()} for a character scalar constrained to a fixed set of allowed values.
}
\examples{
type_prop <- enum(c("string", "number", "boolean"), default = "string")
}
\author{
EDG
}
