% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/msg.R
\name{get_msg_sink}
\alias{get_msg_sink}
\title{Get the current rtemis message sink}
\usage{
get_msg_sink()
}
\value{
The currently registered sink function, or \code{NULL} if none is set.
}
\description{
Get the current rtemis message sink
}
\seealso{
\code{\link[=set_msg_sink]{set_msg_sink()}}, \code{\link[=with_msg_sink]{with_msg_sink()}}.
}
\author{
EDG
}
