% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_strings.R
\name{highlightbig}
\alias{highlightbig}
\title{Highlight big numbers}
\usage{
highlightbig(x, output_type = c("ansi", "html", "plain"))
}
\arguments{
\item{x}{Numeric: Input}

\item{output_type}{Character: Output type ("ansi", "html", "plain").}
}
\value{
Character: Formatted number with thousands separators and highlighting.
}
\description{
Highlight big numbers
}
\examples{
message(highlightbig(1234567))
}
\keyword{internal}
