% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/html.R
\name{html_strong}
\alias{html_strong}
\title{Build a \code{strong} element}
\usage{
html_strong(..., class = NULL, style = NULL, id = NULL)
}
\arguments{
\item{...}{Children: Character, numbers, \code{rtemis_html} objects, or lists of
those.}

\item{class}{Optional Character: \code{class} attribute.}

\item{style}{Optional Character: \code{style} attribute.}

\item{id}{Optional Character: \code{id} attribute.}
}
\value{
Character of class \code{rtemis_html_element} and \code{rtemis_html}.
}
\description{
Build a \code{strong} element
}
\examples{
html_strong(42L)
}
\author{
EDG
}
