% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/00_S7_properties.R
\name{integer_scalar}
\alias{integer_scalar}
\title{Integer scalar S7 property}
\usage{
integer_scalar
}
\value{
An S7 property object.
}
\description{
S7 property accepting a single non-NA integer value (must be \code{integer} type, e.g. \code{1L}).
}
\author{
EDG
}
