% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/03_S7_properties.R
\name{nonneg_double_scalar}
\alias{nonneg_double_scalar}
\title{Non-negative double scalar S7 property}
\usage{
nonneg_double_scalar
}
\value{
An S7 property object.
}
\description{
S7 property accepting a single finite double greater than or equal to zero, i.e. in \eqn{[0, \infty)}.
}
\author{
EDG
}
