% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/03_S7_properties.R
\name{nonneg_integer_scalar}
\alias{nonneg_integer_scalar}
\title{Non-negative integer scalar S7 property}
\usage{
nonneg_integer_scalar
}
\value{
An S7 property object.
}
\description{
S7 property accepting a single non-NA integer value greater than or equal to zero,
i.e. in \eqn{[0, \infty)} (e.g. \code{0L}, \code{1L}).
}
\examples{
Counter <- S7::new_class(
  "Counter",
  properties = list(n_failed = nonneg_integer_scalar)
)
Counter(n_failed = 0L)@n_failed
try(Counter(n_failed = -1L))
}
\author{
EDG
}
