% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/00_S7_properties.R
\name{nonneg_integer_scalar}
\alias{nonneg_integer_scalar}
\title{Non-negative integer scalar S7 property}
\usage{
nonneg_integer_scalar
}
\value{
An S7 property object.
}
\description{
S7 property accepting a single non-NA integer value greater than or equal to zero,
i.e. in \eqn{[0, \infty)} (e.g. \code{0L}, \code{1L}).
}
\author{
EDG
}
