% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_strings.R
\name{objcat}
\alias{objcat}
\title{Cat object}
\usage{
objcat(
  x,
  col = col_object,
  pad = 0L,
  prefix = NULL,
  output_type = c("ansi", "html", "plain")
)
}
\arguments{
\item{x}{Character: Object description}

\item{col}{Character: Color code for the object name}

\item{pad}{Integer: Number of spaces to pad the message with.}

\item{prefix}{Optional Character: Prefix to add before the object name.}

\item{output_type}{Character: Output type ("ansi", "html", "plain").}
}
\value{
NULL: Prints the formatted object description to the console.
}
\description{
Cat object
}
\examples{
objcat("Supervised")
}
\author{
EDG
}
\keyword{internal}
