% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/types.R
\name{optional}
\alias{optional}
\title{Create an optional S7 type}
\usage{
optional(type)
}
\arguments{
\item{type}{S7 base class or S7 class.}
}
\value{
An S7 union type that allows for the specified type or \code{NULL}.
}
\description{
Creates an S7 union type that allows for the specified type or \code{NULL}.
}
\examples{
# Create an optional character type
optional(S7::class_character)
}
\author{
EDG
}
