% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/03_S7_properties.R
\name{optional_character_scalar}
\alias{optional_character_scalar}
\title{Optional non-empty character scalar S7 property}
\usage{
optional_character_scalar
}
\value{
An S7 property object.
}
\description{
S7 property accepting \code{NULL} or a single non-NA, non-empty (after trimming whitespace) string.
}
\author{
EDG
}
