% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/00_S7_properties.R
\docType{data}
\name{optional_character_scalar}
\alias{optional_character_scalar}
\title{Optional non-empty character scalar S7 property}
\format{
An object of class \code{S7_property} of length 6.
}
\usage{
optional_character_scalar
}
\value{
An S7 property object.
}
\description{
S7 property accepting \code{NULL} or a single non-NA, non-empty (after trimming whitespace) string.
}
\author{
EDG
}
\keyword{datasets}
