% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/00_S7_properties.R
\name{optional_double_scalar}
\alias{optional_double_scalar}
\title{Optional double scalar S7 property}
\usage{
optional_double_scalar
}
\value{
An S7 property object.
}
\description{
S7 property accepting \code{NULL} or a single non-NA double value.
}
\author{
EDG
}
