% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/03_S7_properties.R
\name{optional_integer_scalar}
\alias{optional_integer_scalar}
\title{Optional integer scalar S7 property}
\usage{
optional_integer_scalar
}
\value{
An S7 property object.
}
\description{
S7 property accepting \code{NULL} or a single non-NA integer value (must be \code{integer} type, e.g. \code{1L}).
}
\author{
EDG
}
