% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/03_S7_properties.R
\name{optional_logical_scalar}
\alias{optional_logical_scalar}
\title{Optional logical scalar S7 property}
\usage{
optional_logical_scalar
}
\value{
An S7 property object.
}
\description{
S7 property accepting \code{NULL} or a single non-NA logical value.
}
\author{
EDG
}
