% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/00_S7_properties.R
\name{optional_nonneg_integer_scalar}
\alias{optional_nonneg_integer_scalar}
\title{Optional non-negative integer scalar S7 property}
\usage{
optional_nonneg_integer_scalar
}
\value{
An S7 property object.
}
\description{
S7 property accepting \code{NULL} or a single non-NA integer value greater than or equal to zero.
}
\author{
EDG
}
