% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/00_S7_properties.R
\name{optional_pos_double_vector}
\alias{optional_pos_double_vector}
\title{Optional positive double vector S7 property}
\usage{
optional_pos_double_vector
}
\value{
An S7 property object.
}
\description{
S7 property accepting \code{NULL} or a non-empty double vector with all elements finite,
strictly greater than zero, and no NAs.
}
\author{
EDG
}
