% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/00_S7_properties.R
\docType{data}
\name{optional_pos_integer_scalar}
\alias{optional_pos_integer_scalar}
\title{Optional positive integer scalar S7 property}
\format{
An object of class \code{S7_property} of length 6.
}
\usage{
optional_pos_integer_scalar
}
\value{
An S7 property object.
}
\description{
S7 property accepting \code{NULL} or a single non-NA integer value strictly greater than zero.
}
\author{
EDG
}
\keyword{datasets}
