% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/00_S7_properties.R
\name{optional_prob_scalar}
\alias{optional_prob_scalar}
\title{Optional probability scalar S7 property}
\usage{
optional_prob_scalar
}
\value{
An S7 property object.
}
\description{
S7 property accepting \code{NULL} or a single finite double in \eqn{[0, 1]}.
}
\author{
EDG
}
