% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/00_S7_properties.R
\docType{data}
\name{pos_double_scalar}
\alias{pos_double_scalar}
\title{Positive double scalar S7 property}
\format{
An object of class \code{S7_property} of length 6.
}
\usage{
pos_double_scalar
}
\value{
An S7 property object.
}
\description{
S7 property accepting a single finite double strictly greater than zero, i.e. in \eqn{(0, \infty)}.
}
\author{
EDG
}
\keyword{datasets}
