% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/03_S7_properties.R
\name{pos_integer_scalar}
\alias{pos_integer_scalar}
\title{Positive integer scalar S7 property}
\usage{
pos_integer_scalar
}
\value{
An S7 property object.
}
\description{
S7 property accepting a single non-NA integer value strictly greater than zero (e.g. \code{1L}).
}
\examples{
Resample <- S7::new_class(
  "Resample",
  properties = list(n_resamples = pos_integer_scalar)
)
Resample(n_resamples = 10L)@n_resamples
try(Resample(n_resamples = 0L))
}
\author{
EDG
}
