% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/00_S7_properties.R
\docType{data}
\name{pos_integer_scalar}
\alias{pos_integer_scalar}
\title{Positive integer scalar S7 property}
\format{
An object of class \code{S7_property} of length 6.
}
\usage{
pos_integer_scalar
}
\value{
An S7 property object.
}
\description{
S7 property accepting a single non-NA integer value strictly greater than zero (e.g. \code{1L}).
}
\author{
EDG
}
\keyword{datasets}
