% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/03_S7_properties.R
\name{prob_scalar}
\alias{prob_scalar}
\title{Probability scalar S7 property}
\usage{
prob_scalar
}
\value{
An S7 property object.
}
\description{
S7 property accepting a single finite double in \eqn{[0, 1]}.
}
\examples{
Split <- S7::new_class("Split", properties = list(train_p = prob_scalar))
Split(train_p = 0.75)@train_p
try(Split(train_p = 1.5))
}
\author{
EDG
}
