% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/00_S7_properties.R
\docType{data}
\name{prob_scalar}
\alias{prob_scalar}
\title{Probability scalar S7 property}
\format{
An object of class \code{S7_property} of length 6.
}
\usage{
prob_scalar
}
\value{
An S7 property object.
}
\description{
S7 property accepting a single finite double in \eqn{[0, 1]}.
}
\author{
EDG
}
\keyword{datasets}
