% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/03_S7_properties.R
\name{prob_vector}
\alias{prob_vector}
\title{Probability vector S7 property}
\usage{
prob_vector
}
\value{
An S7 property object.
}
\description{
S7 property accepting a non-empty double vector with all elements in \eqn{[0, 1]} and no NAs.
}
\examples{
Preds <- S7::new_class("Preds", properties = list(probabilities = prob_vector))
Preds(probabilities = c(0, 0.5, 1))@probabilities
try(Preds(probabilities = c(0.5, NA)))
}
\author{
EDG
}
