% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rtemis_color_system.R
\name{rtemis_colors}
\alias{rtemis_colors}
\title{rtemis Colors}
\usage{
rtemis_colors
}
\value{
Named character vector of hex color codes.
}
\description{
A named vector of colors used in the rtemis ecosystem, provided as hex strings.
}
\examples{
rtemis_colors[["teal"]]
}
\author{
EDG
}
