% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_print.R
\name{show_df}
\alias{show_df}
\title{Show data.frame}
\usage{
show_df(
  x,
  pad = 0L,
  spacing = 1L,
  ddSci_dp = NULL,
  transpose = FALSE,
  justify = "right",
  incl_colnames = TRUE,
  incl_rownames = TRUE,
  colnames_formatter = highlight,
  rownames_formatter = gray,
  output_type = c("ansi", "html", "plain")
)
}
\arguments{
\item{x}{data frame}

\item{pad}{Integer: Pad output with this many spaces.}

\item{spacing}{Integer: Number of spaces between columns.}

\item{ddSci_dp}{Integer: Number of decimal places to print using \link{ddSci}. Default = NULL for no
formatting}

\item{transpose}{Logical: If TRUE, transpose \code{x} before printing.}

\item{justify}{Character: "right", "left".}

\item{incl_colnames}{Logical: If TRUE, include column names.}

\item{incl_rownames}{Logical: If TRUE, include row names.}

\item{colnames_formatter}{Format function for printing column names.}

\item{rownames_formatter}{Format function for printing row names.}

\item{output_type}{Character: Output type ("ansi", "html", "plain").}
}
\value{
Character: Formatted string representation of the data.frame.
}
\description{
Create a pretty text representation of a data.frame.
}
\examples{
show_df(iris[1:3, ]) |> cat()
}
\author{
EDG
}
\keyword{internal}
