% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/00_S7_properties.R
\name{unit_open_scalar}
\alias{unit_open_scalar}
\title{Open-unit-interval scalar S7 property}
\usage{
unit_open_scalar
}
\value{
An S7 property object.
}
\description{
S7 property accepting a single finite double strictly in \eqn{(0, 1)}.
}
\author{
EDG
}
