% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/03_S7_properties.R
\name{unit_open_vector}
\alias{unit_open_vector}
\title{Open-unit-interval vector S7 property}
\usage{
unit_open_vector
}
\value{
An S7 property object.
}
\description{
S7 property accepting a non-empty double vector with all elements strictly in \eqn{(0, 1)}
and no NAs.
}
\examples{
Grid <- S7::new_class("Grid", properties = list(learning_rates = unit_open_vector))
Grid(learning_rates = c(0.01, 0.1))@learning_rates
try(Grid(learning_rates = c(0.1, 1)))
}
\author{
EDG
}
