% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/00_S7_properties.R
\docType{data}
\name{unit_open_vector}
\alias{unit_open_vector}
\title{Open-unit-interval vector S7 property}
\format{
An object of class \code{S7_property} of length 6.
}
\usage{
unit_open_vector
}
\value{
An S7 property object.
}
\description{
S7 property accepting a non-empty double vector with all elements strictly in \eqn{(0, 1)}
and no NAs.
}
\author{
EDG
}
\keyword{datasets}
